pyproject.toml
[tool.poetry]
name = "projeto-ml-senai"
version = "0.1.0"
description = ""
authors = ["venysssssssssss <shablaur7@hotmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.2.2"
numpy = "^2.0.1"
scikit-learn = "^1.5.1"
faker = "^26.2.0"
joblib = "^1.4.2"
matplotlib = "^3.9.1.post1"
pyyaml = "^6.0.2"
seaborn = "^0.13.2"
mkdocs = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
